css/css-flexbox/flex-minimum-height-flex-items-025.html

<!DOCTYPE html>
<link rel="author" title="dgrogan@chromium.org" href="mailto:dgrogan@chromium.org" />
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#min-size-auto" />
<link rel="bookmark" href="https://crbug.com/1033476">
<meta name="assert" content="Column flex item's height does not influence its content size suggestion in a varitey of scenarios.">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>

<style>
.outer-column-flexbox {
  display: flex;
  flex-direction: column;
  width: 100px;
  margin-bottom: 10px; /* Just for visually spacing things out */
}
</style>

<body onload="checkLayout('.outer-column-flexbox > div, .row-flex-item');">

<div id=log></div>

<p>Test passes if there is a sequence of filled green squares.</p>

<div class=outer-column-flexbox>
  <div style="flex-basis: 100px; height: 200px;" data-expected-height=100>
    <div style="height: 100%; background: green;" data-expected-height=100></div>
  </div>
</div>

<div class=outer-column-flexbox>
  <div style="flex-basis: 100px; height: 200px; writing-mode: vertical-rl;" data-expected-height=100>
    <div style="height: 100%; background: green; width: 100px;" data-expected-height=100></div>
  </div>
</div>

<div class=outer-column-flexbox>
  <div style="flex-basis: 100px; height: 200px;" data-expected-height=100>
    <div style="display: grid; height: 100%; background: green" data-expected-height=100></div>
  </div>
</div>

<div class=outer-column-flexbox>
  <div style="flex-basis: 100px; height: 200px;" data-expected-height=100>
    <table data-expected-height=100 style="height: 100%; width: 100%; background: green">
      <td>
        <div style="height: 50px;"></div>
      </td>
    </table>
  </div>
</div>

<div style="display: flex; width: 0px; height: 20px;">
  <div style="flex-basis: 50px; height: 100px;" data-expected-width=100 class=row-flex-item>
    <img src="support/200x200-green.png" style="height: 100%">
  </div>
</div>
